+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_factory.py
@@ -3,15 +3,6 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.mgmt.storage import StorageManagementClient
-
-from azure.storage import CloudStorageAccount
-from azure.storage.blob import BlockBlobService
-from azure.storage.file import FileService
-from azure.storage.table import TableService
-from azure.storage.queue import QueueService
-from azure.storage._error import _ERROR_STORAGE_MISSING_INFO
-
 from azure.cli.core.commands.client_factory import get_mgmt_service_client, get_data_service_client
 from azure.cli.core.commands import CLIError
 from azure.cli.core._profile import CLOUD
@@ -42,6 +33,7 @@ def generic_data_service_factory(service, name=None, key=None, connection_string
     try:
         return get_storage_data_service_client(service, name, key, connection_string, sas_token)
     except ValueError as val_exception:
+        from azure.storage._error import _ERROR_STORAGE_MISSING_INFO
         message = str(val_exception)
         if message == _ERROR_STORAGE_MISSING_INFO:
             message = NO_CREDENTIALS_ERROR_MESSAGE
@@ -49,10 +41,12 @@ def generic_data_service_factory(service, name=None, key=None, connection_string
 
 
 def storage_client_factory(**_):
+    from azure.mgmt.storage import StorageManagementClient
     return get_mgmt_service_client(StorageManagementClient)
 
 
 def file_data_service_factory(kwargs):
+    from azure.storage.file import FileService
     return generic_data_service_factory(
         FileService,
         kwargs.pop('account_name', None),
@@ -62,6 +56,7 @@ def file_data_service_factory(kwargs):
 
 
 def blob_data_service_factory(kwargs):
+    from azure.storage.blob import BlockBlobService
     from ._params import blob_types
     blob_type = kwargs.get('blob_type')
     blob_service = blob_types.get(blob_type, BlockBlobService)
@@ -74,6 +69,7 @@ def blob_data_service_factory(kwargs):
 
 
 def table_data_service_factory(kwargs):
+    from azure.storage.table import TableService
     return generic_data_service_factory(
         TableService,
         kwargs.pop('account_name', None),
@@ -83,6 +79,7 @@ def table_data_service_factory(kwargs):
 
 
 def queue_data_service_factory(kwargs):
+    from azure.storage.queue import QueueService
     return generic_data_service_factory(
         QueueService,
         kwargs.pop('account_name', None),
@@ -92,6 +89,7 @@ def queue_data_service_factory(kwargs):
 
 
 def cloud_storage_account_service_factory(kwargs):
+    from azure.storage import CloudStorageAccount
     account_name = kwargs.pop('account_name', None)
     account_key = kwargs.pop('account_key', None)
     sas_token = kwargs.pop('sas_token', None)
