+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_validators.py
@@ -6,17 +6,13 @@
 import argparse
 import os
 import re
-from collections import OrderedDict
 from datetime import datetime, timedelta
 
-from azure.cli.core._config import az_config
-from azure.cli.core._profile import CLOUD
 from azure.cli.core.util import CLIError
-from azure.cli.core.commands.client_factory import get_mgmt_service_client
-from azure.cli.core.commands.validators import validate_key_value_pairs
-from azure.mgmt.storage import StorageManagementClient
-from azure.storage.sharedaccesssignature import SharedAccessSignature
-from azure.storage.blob import Include, PublicAccess
+from azure.cli.core._profile import CLOUD
+
+
+from azure.storage.blob import PublicAccess
 from azure.storage.blob.baseblobservice import BaseBlobService
 from azure.storage.blob.models import ContentSettings as BlobContentSettings, BlobPermissions
 from azure.storage.file import FileService
@@ -26,24 +22,15 @@ from azure.storage.table import TablePermissions, TablePayloadFormat
 
 from ._factory import get_storage_data_service_client
 from .util import glob_files_locally
+from ._command_type import validate_client_parameters, query_account_key
 
 storage_account_key_options = {'primary': 'key1', 'secondary': 'key2'}
 
 
 # Utilities
 
-def _query_account_key(account_name):
-    scf = get_mgmt_service_client(StorageManagementClient)
-    acc = next((x for x in scf.storage_accounts.list() if x.name == account_name), None)
-    if acc:
-        from azure.cli.core.commands.arm import parse_resource_id
-        rg = parse_resource_id(acc.id)['resource_group']
-        return scf.storage_accounts.list_keys(rg, account_name).keys[0].value  # pylint: disable=no-member
-    else:
-        raise ValueError("Storage account '{}' not found.".format(account_name))
-
-
 def _create_short_lived_blob_sas(account_name, account_key, container, blob):
+    from azure.storage.sharedaccesssignature import SharedAccessSignature
     expiry = (datetime.utcnow() + timedelta(days=1)).strftime('%Y-%m-%dT%H:%M:%SZ')
     sas = SharedAccessSignature(account_name, account_key)
     return sas.generate_blob(container, blob, permission=BlobPermissions(read=True), expiry=expiry,
@@ -51,6 +38,7 @@ def _create_short_lived_blob_sas(account_name, account_key, container, blob):
 
 
 def _create_short_lived_file_sas(account_name, account_key, share, directory_name, file_name):
+    from azure.storage.sharedaccesssignature import SharedAccessSignature
     # if dir is empty string change it to None
     directory_name = directory_name if directory_name else None
     expiry = (datetime.utcnow() + timedelta(days=1)).strftime('%Y-%m-%dT%H:%M:%SZ')
@@ -70,37 +58,6 @@ def validate_accept(namespace):
         }
         namespace.accept = formats[namespace.accept.lower()]
 
-
-def validate_client_parameters(namespace):
-    """ Retrieves storage connection parameters from environment variables and parses out
-    connection string into account name and key """
-    n = namespace
-
-    if not n.connection_string:
-        n.connection_string = az_config.get('storage', 'connection_string', None)
-
-    # if connection string supplied or in environment variables, extract account key and name
-    if n.connection_string:
-        conn_dict = validate_key_value_pairs(n.connection_string)
-        n.account_name = conn_dict['AccountName']
-        n.account_key = conn_dict['AccountKey']
-
-    # otherwise, simply try to retrieve the remaining variables from environment variables
-    if not n.account_name:
-        n.account_name = az_config.get('storage', 'account', None)
-    if not n.account_key:
-        n.account_key = az_config.get('storage', 'key', None)
-    if not n.sas_token:
-        n.sas_token = az_config.get('storage', 'sas_token', None)
-
-    # strip the '?' from sas token. the portal and command line are returns sas token in different
-    # forms
-    if n.sas_token:
-        n.sas_token = n.sas_token.lstrip('?')
-
-    # if account name is specified but no key, attempt to query
-    if n.account_name and not n.account_key and not n.sas_token:
-        n.account_key = _query_account_key(n.account_name)
 
 
 def validate_source_uri(namespace):  # pylint: disable=too-many-statements
@@ -171,7 +128,7 @@ def validate_source_uri(namespace):  # pylint: disable=too-many-statements
             # the source account is different from destination account but the key is missing
             # try to query one.
             try:
-                source_account_key = _query_account_key(source_account_name)
+                source_account_key = query_account_key(source_account_name)
             except ValueError:
                 raise ValueError('Source storage account {} not found.'.format(source_account_name))
     # else: both source account name and key are given by user
@@ -352,6 +309,7 @@ def validate_included_datasets(namespace):
         if set(include) - set('cms'):
             help_string = '(c)opy-info (m)etadata (s)napshots'
             raise ValueError('valid values are {} or a combination thereof.'.format(help_string))
+        from azure.storage.blob import Include
         namespace.include = Include('s' in include, 'm' in include, False, 'c' in include)
 
 
@@ -475,7 +433,7 @@ def get_source_file_or_blob_service_client(namespace):
         if not (source_key or source_sas):
             # when either storage account key or SAS is given, try to fetch the key in the current
             # subscription
-            source_key = _query_account_key(source_account)
+            source_key = query_account_key(source_account)
 
         if source_container:
             from azure.storage.blob.blockblobservice import BlockBlobService
@@ -709,108 +667,5 @@ def services_type(string):
     if set(string) - set("bqtf"):
         raise ValueError
     return Services(_str=''.join(set(string)))
-
-# endregion
-
-# region TRANSFORMS
-
-
-def transform_acl_list_output(result):
-    """ Transform to convert SDK output into a form that is more readily
-    usable by the CLI and tools such as jpterm. """
-    new_result = []
-    for key in sorted(result.keys()):
-        new_entry = OrderedDict()
-        new_entry['Name'] = key
-        new_entry['Start'] = result[key]['start']
-        new_entry['Expiry'] = result[key]['expiry']
-        new_entry['Permissions'] = result[key]['permission']
-        new_result.append(new_entry)
-    return new_result
-
-
-def transform_container_permission_output(result):
-    return {'publicAccess': result.public_access or 'off'}
-
-
-def transform_cors_list_output(result):
-    new_result = []
-    for service in sorted(result.keys()):
-        service_name = service
-        for i, rule in enumerate(result[service]):
-            new_entry = OrderedDict()
-            new_entry['Service'] = service_name
-            service_name = ''
-            new_entry['Rule'] = i + 1
-
-            new_entry['AllowedMethods'] = ', '.join((x for x in rule.allowed_methods))
-            new_entry['AllowedOrigins'] = ', '.join((x for x in rule.allowed_origins))
-            new_entry['ExposedHeaders'] = ', '.join((x for x in rule.exposed_headers))
-            new_entry['AllowedHeaders'] = ', '.join((x for x in rule.allowed_headers))
-            new_entry['MaxAgeInSeconds'] = rule.max_age_in_seconds
-            new_result.append(new_entry)
-    return new_result
-
-
-def transform_entity_query_output(result):
-    new_results = []
-    ignored_keys = ['etag', 'Timestamp', 'RowKey', 'PartitionKey']
-    for row in result['items']:
-        new_entry = OrderedDict()
-        new_entry['PartitionKey'] = row['PartitionKey']
-        new_entry['RowKey'] = row['RowKey']
-        other_keys = sorted([x for x in row.keys() if x not in ignored_keys])
-        for key in other_keys:
-            new_entry[key] = row[key]
-        new_results.append(new_entry)
-    return new_results
-
-
-def transform_logging_list_output(result):
-    new_result = []
-    for key in sorted(result.keys()):
-        new_entry = OrderedDict()
-        new_entry['Service'] = key
-        new_entry['Read'] = str(result[key]['read'])
-        new_entry['Write'] = str(result[key]['write'])
-        new_entry['Delete'] = str(result[key]['delete'])
-        new_entry['RetentionPolicy'] = str(result[key]['retentionPolicy']['days'])
-        new_result.append(new_entry)
-    return new_result
-
-
-def transform_metrics_list_output(result):
-    new_result = []
-    for service in sorted(result.keys()):
-        service_name = service
-        for interval in sorted(result[service].keys()):
-            item = result[service][interval]
-            new_entry = OrderedDict()
-            new_entry['Service'] = service_name
-            service_name = ''
-            new_entry['Interval'] = str(interval)
-            new_entry['Enabled'] = str(item['enabled'])
-            new_entry['IncludeApis'] = str(item['includeApis'])
-            new_entry['RetentionPolicy'] = str(item['retentionPolicy']['days'])
-            new_result.append(new_entry)
-    return new_result
-
-
-def create_boolean_result_output_transformer(property_name):
-    def _transformer(result):
-        return {property_name: result}
-
-    return _transformer
-
-
-def transform_storage_list_output(result):
-    return list(result)
-
-
-def transform_url(result):
-    """ Ensures the resulting URL string does not contain extra / characters """
-    result = re.sub('//', '/', result)
-    result = re.sub('/', '//', result, count=1)
-    return result
 
 # endregion
